requirements.txt
numba==0.56.4
numpy==1.23.5
pandas>=2.0.2
matplotlib>=3.7.1
openpyxl>=3.1.2
git+https://github.com/MaartenGr/BERTopic.git#egg=bertopic
hdbscan>=0.8.29
umap-learn>=0.5.3
sentence-transformers>=2.2.2
